+++ vo06/fhooe-router-skeleton/CHANGELOG.md
@@ -14,6 +14,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Security
 
+## [3.0.1] - 2026-05-10
+
+### Changed
+
+- Updated `latte/latte` to 3.1.0, also dev dependencies.
+
+### Fixed
+
+- Replaced deprecated Latte setTempDirectory() with setCacheDirectory() in index.php.
+
+## [3.0.0] - 2026-03-24
+
+### Changed
+
+- Based on `Router` 3.0.0.
+- Switched from `fhooe/twig-extensions` to `fhooe/latte-extensions` (`RouterExtension`, `SessionExtension`).
+- Switched all templates from Twig (`.twig`) to Latte (`.latte`).
+- `$router->basePath` is now set as a property assignment instead of `setBasePath()`.
+- Switched to PHP 8.5 as a minimum requirement.
+- Updated all dependencies.
+
+### Removed
+
+- Removed `twig/twig` and `fhooe/twig-extensions` dependencies.
+
 ## [2.0.0] - 2025-03-26
 
 ### Added
@@ -136,7 +161,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added notes on Contributing.
 - Added this changelog.
 
-[Unreleased]: https://github.com/Digital-Media/fhooe-router-skeleton/compare/v2.0.0...HEAD
+[Unreleased]: https://github.com/Digital-Media/fhooe-router-skeleton/compare/v3.0.1...HEAD
+[3.0.1]: https://github.com/Digital-Media/fhooe-router-skeleton/compare/v3.0.0...v3.0.1
+[3.0.0]: https://github.com/Digital-Media/fhooe-router-skeleton/compare/v2.0.0...v3.0.0
 [2.0.0]: https://github.com/Digital-Media/fhooe-router-skeleton/compare/v1.0.1...v2.0.0
 [1.0.1]: https://github.com/Digital-Media/fhooe-router-skeleton/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/Digital-Media/fhooe-router-skeleton/compare/v0.3.1...v1.0.0